flashcards/backend/pyproject.toml
[tool.poetry]
name = "flashcards"
version = "0.1.0"
description = "virtual environment for flashcards application"
authors = ["Azeno Team"]

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
dj-database-url = "1.2.0"
gunicorn = "20.1.0"
django-debug-toolbar = "4.0.0"
pytest = "7.2.2"
pytest-django = "4.5.2"
coverage = "7.2.1"
pytest-cov = "4.0.0"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.2.2"
python-magic = "0.4.27"
pillow = "9.4.0"
psycopg = "3.1.8"
django-filter = "23.1"
types-factory-boy = "0.4.1"

[tool.poetry.group.dev.dependencies]
sphinx = "6.1.3"
renku-sphinx-theme = "0.2.1"
sphinx-rtd-theme = "1.2.0"
black = "23.1.0"
mypy = "^1.1.1"
factory-boy = "3.2.1"
django-stubs = "1.15.0"
djangorestframework-stubs = "1.9.1"
types-dj-database-url = "1.2.0.1"
types-factory-boy = "0.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = "test_*.py"

addopts = [
    "--cov=.",
    "--cov-config=.coveragerc",
    "--cov-report",
    "term-missing:skip-covered",
    "--cov-fail-under=80"
]

[tool.pytest.markers]
slow = "Run slow unit test e.g. for models"
unit = "Run unit tests"
smoke = "Run core unit tests"
dbtest = "Run database tests"


[tool.black]
line-length = 120
target-version = ['py311']
exclude = '''
/(
    | \.git/  # exclude a directory
    | \.mypy_cache/  # exclude a directory
    | \.pytest_cache/  # exclude a directory
    | .*\.pyi?  # exclude generated files
    | .*\.json  # exclude metadata files
    | .*\.md  # exclude documentation
    | .*\.toml  # exclude configuration files
    | .*\.yml  # exclude configuration files
    | .*\.yaml  # exclude configuration files
    | .*\.txt  # exclude configuration files
    | .*\.lock  # exclude configuration files
    | .*\.sh  # exclude configuration files
    | .*\.log  # exclude configuration files
)/
'''
